test vectors: move non-numeric Ordering instances to test suite
