+++ sigmastate/src/main/scala/sigmastate/eval/BigIntegerOps.scala
@@ -21,98 +21,6 @@ object OrderingOps {
     def compare(x: BigInt, y: BigInt) = x.compareTo(y)
   }
   implicit object BigIntOrdering extends BigIntOrdering
-
-  trait GroupElementOrdering extends Ordering[GroupElement] {
-    /** Compares `x: ECPoint` string representation with `y: ECPoint` string for order.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: GroupElement, y: GroupElement) = {
-      SigmaDsl.toECPoint(x).toString.compareTo(SigmaDsl.toECPoint(y).toString)
-    }
-  }
-  implicit object GroupElementOrdering extends GroupElementOrdering
-
-  trait AvlTreeOrdering extends Ordering[AvlTree] {
-    /** Compares this `x: AvlTree` string representation with `y: AvlTree` string for order.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: AvlTree, y: AvlTree) = {
-      x.toString.compareTo(y.toString)
-    }
-  }
-  implicit object AvlTreeOrdering extends AvlTreeOrdering
-
-  class CollOrdering[T: Ordering] extends Ordering[Coll[T]] {
-    implicit val O = implicitly[Ordering[Iterable[T]]]
-
-    /** Compares this `x: Coll` with `y: Coll` using Ordering for underlying Array.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: Coll[T], y: Coll[T]) = {
-      O.compare(x.toArray, y.toArray)
-    }
-  }
-  implicit def collOrdering[T: Ordering]: Ordering[Coll[T]] = new CollOrdering[T]
-
-  trait BoxOrdering extends Ordering[Box] {
-    /** Compares this `x: Box` string representation with `y: Box` string for order.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: Box, y: Box) = {
-      x.toString.compareTo(y.toString)
-    }
-  }
-  implicit object BoxOrdering extends BoxOrdering
-
-  trait PreHeaderOrdering extends Ordering[PreHeader] {
-    /** Compares this `x: PreHeader` with `y: PreHeader` using block height.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: PreHeader, y: PreHeader) = {
-      Ordering.Int.compare(x.height, y.height)
-    }
-  }
-  implicit object PreHeaderOrdering extends PreHeaderOrdering
-
-  trait HeaderOrdering extends Ordering[Header] {
-    /** Compares this `x: Header` with `y: Header` using block height.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: Header, y: Header) = {
-      Ordering.Int.compare(x.height, y.height)
-    }
-  }
-  implicit object HeaderOrdering extends HeaderOrdering
-
-  trait ContextOrdering extends Ordering[Context] {
-    val O = Ordering[(Int, Coll[Byte])]
-
-    /** Compares this `x: Context` with `y: Context` using block height and SELF.id.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: Context, y: Context) = {
-      O.compare((x.HEIGHT, x.SELF.id), (y.HEIGHT, y.SELF.id))
-    }
-  }
-  implicit object ContextOrdering extends ContextOrdering
-
-  trait SigmaPropOrdering extends Ordering[SigmaProp] {
-    /** Compares this `x: SigmaProp` with `y: SigmaProp` using string representation.
-      * @returns a negative integer, zero, or a positive integer as the
-      * `x` is less than, equal to, or greater than `y`.
-      */
-    def compare(x: SigmaProp, y: SigmaProp) = {
-      x.toString.compareTo(y.toString)
-    }
-  }
-  implicit object SigmaPropOrdering extends SigmaPropOrdering
 }
 
 object NumericOps {
